add schemas, contact-service refactoring
