+++ models/contactsService/contact-service.js
@@ -1,51 +1,23 @@
-const { nanoid } = require("nanoid");
-const { Contact } = require("../../server");
-
-const writeContacts = async (contacts) => {
-  await fs.writeFile(contactsPath, JSON.stringify(contacts, null, 2));
-};
+const { Contact } = require("../../schemas/contact-db-schema");
 
 const listContacts = async () => {
-  const contacts = await Contact.find();
-  return JSON.parse(contacts);
+  return await Contact.find();
 };
 
 const getContactById = async (contactId) => {
-  const contacts = await listContacts();
-  return contacts.find((item) => item.id === contactId) || null;
+  return await Contact.findById(contactId);
 };
 
 const removeContact = async (contactId) => {
-  const contacts = await listContacts();
-  const index = contacts.findIndex((item) => item.id === contactId);
-  if (index === -1) {
-    return null;
-  }
-  const [result] = contacts.splice(index, 1);
-  await writeContacts(contacts);
-  return result;
+  return await Contact.findOneAndRemove(contactId);
 };
 
 const addContact = async (body) => {
-  const contacts = await listContacts();
-  const newContact = {
-    id: nanoid(),
-    ...body,
-  };
-  contacts.push(newContact);
-  await writeContacts(contacts);
-  return newContact;
+  return await Contact.create(body);
 };
 
-const updateContact = async (id, body) => {
-  const contacts = await listContacts();
-  const index = contacts.findIndex((item) => item.id === id);
-  if (index === -1) {
-    return null;
-  }
-  contacts[index] = { id, ...body };
-  await writeContacts(contacts);
-  return contacts[index];
+const updateContact = async (contactId, body) => {
+  return await Contact.findByIdAndUpdate(contactId, body);
 };
 
 module.exports = {
