+++ routes/api/contacts-routes.js
@@ -1,24 +1,20 @@
 const express = require("express");
 const contactsController = require("../../controllers/contacts-controller");
 const router = express.Router();
-const schemas = require("../../schemas/contacts-schemas");
+const contactSchema = require("../../schemas/contacts-api-schema");
 const { validateBody } = require("../../decorators/validateBody");
 
 router.get("/", contactsController.listContacts);
 
 router.get("/:contactId", contactsController.getContactById);
 
-router.post(
-  "/",
-  validateBody(schemas.contactSchema),
-  contactsController.addContact
-);
+router.post("/", validateBody(contactSchema), contactsController.addContact);
 
 router.delete("/:contactId", contactsController.removeContact);
 
 router.put(
   "/:contactId",
-  validateBody(schemas.contactSchema),
+  validateBody(contactSchema),
   contactsController.updateContact
 );
