+++ server.js
@@ -1,26 +1,8 @@
-const app = require("./app");
 const mongoose = require("mongoose");
 
+const app = require("./app");
+
 const { DB_HOST } = process.env;
-
-const contactSchema = new mongoose.Schema({
-  name: {
-    type: String,
-    required: [true, "Set name for contact"],
-  },
-  email: {
-    type: String,
-  },
-  phone: {
-    type: String,
-  },
-  favorite: {
-    type: Boolean,
-    default: false,
-  },
-});
-
-const Contact = mongoose.model("contacts", contactSchema);
 
 async function dbConnect() {
   try {
@@ -35,5 +17,3 @@ async function dbConnect() {
 }
 
 dbConnect();
-
-module.exports = { Contact };
